Add skeleton Vue app in frontend/
Everything appears to work, brunch compiles and serves all assets and
individual directives from components, hot-reloads and builds. I can't
seem to make it use sass's indented syntax[0], so limited to scss for
the time being (perhaps this is actually a more sensible move in the
long run, it appears to be much more popular).

Makefile enhanced to allow frontend to be run with:

```bash
make run-frontend
```

[0] http://sass-lang.com/documentation/file.INDENTED_SYNTAX.html