Add JavaScript CMSFile model, and configure CORS
The CMSFile model will be responsible for all file-releated API
functionality. This version implements the `all(directory)` method,
which returns all file objects that reside in the given `directory`.

To support this in development mode, CORS needs to be accounted for.
This is a config option; in 'live' environments where the Go app serves
everything it won't be needed so should be disabled entirely