Switch Milligram out for Bootstrap
Despite wanting to keep the app as lightweight as possible, my CSS
skills aren't up to writing everything from scratch in a robust,
cross-browser manner. Also, relying on super-new CSS functionality (like
CSS Grid) is *probably* not the best idea; going with something that's
widely used and easy to customise is probably going to be a benefit
while getting this project up-and-running.