Continue building DocumentEdit component
DocumentEdit can now send PATCH requests to the backend containing
changes. There is a bug that means the body isn't stored, which is due
to the backend deserialisation (via `json.Unmarshal`) accepting a
byte slice rather than a string:

> Array and slice values encode as JSON arrays, except that []byte
> encodes as a base64-encoded string, and a nil slice encodes as the
> null JSON value.[0]

It's probably easiest to modify the backend and undo the
`map`/`charCodeAt` stuff in `commit.js`.

[0] https://golang.org/pkg/encoding/json/