Plug FrontMatter into the file editor
Using Particle, the editor now supports the reading and writing of front
matter[0]. Front Matter fields will be displayed on the right of the
editing pane, for the moment along with (unused) Tags and Commit
Message, but they will be made distinct.

To support this, `createFile` had to be split out to `createEmptyFile`,
because blank front matter was added to empty files causing tests to
fail.

[0] https://jekyllrb.com/docs/frontmatter/