Add document creation component and delete button
DocumentNew works similarly to (and is based on) DocumentEdit, and
allows for the creation and writing of front matter and content of a new
document.

For deletion, a 'Delete' button has been added to the 'DocumentShow'
component; clicking it calls the file's `delete` action on the API and, if
successful, redirects the user to the directory index view.