Automatically set the filename and allow override
Documents are accessed by their filename so it makes sense to remove any
difficult-to-deal-with characters beforehand; this is done by José
Quintana's slugify method.

By default, the override default filename box is unchecked and a filename
will be automatically generated based on the title. If it is checked a
user can type in the filename box; the name is slugified on the fly.