Add and integrate SimpleMDE
SimpleMDE[0] is a popular and customisable Markdown editor. It may not
be the final editor used in this project *but* is better than a plain
old `textarea` for the time being.

Rather than just triggering the `CMSFile.find` in the store, returning
the promise allows the document's MarkDown content to by synced when the
promise is resolved

[0] https://simplemde.com/