Begin integration of JWT starting with the backend
Mechanisms for loading and signing/verifying keys have been added
and handlers for creating a signed JWT[0] and using it to protect all
resources that begin with `/api/*`.

Posting the following JSON document to `/auth/login` should, if the
credentials are correct, return a valid JWT which can be used until
it has expired.

```json
{
    "username": "Patty Bouvier",
    "password": "MacGuyver"
}
```

All requests made subsequently to proteceted endpoints must include the
header `Authorization` with value `Bearer JWTTOKEN_GOES_HERE`.

[0] https://jwt.io/