Implement JWT authentication on the front end
Currently very basic and not really integrated with the router, but
functional 🎉

Navigating to `/cms/login` will display a **very basic** login form.
Entering the hard-coded credentials `someone`/`p@ssword` will save the
JWT to localStorage[0] and then we can browse pages of the CMS that
interact with the API properly.

[0] https://developer.mozilla.org/en/docs/Web/API/Window/localStorage