Begin initial setup portion of the API
Starting with the creation of an initial user. Eventually a form will be
displayed instead of the login form **if** there are no users at all.
The first user created will automatically be an administrator (and can
then add further users). This endpoint will not be available if any
users exist

This change also introduces the Validator library[0], which allows for
quick validation rules to be applied to structs. Basic validation error
handling is accounted for and errors are passed back to the API in a
sensible format.

```json
{
    "Username": "is a required field"
}
```

I can't quite see how to make it use the JSON name `username` rather
than the struct attribute name 😣

[0] https://github.com/go-playground/validator