Add API functionality for user management
Add handlers for user management that map to the user specific
functionality in `users.go`. This includes taking the step of making
safe the user data we send out via the API. Passwords are omitted and
all user data visible to 'the outside world' should be in the
`LimitedUser` format.