Add tests for ValidateTokenMiddleware
Make sure that the ValidateTokenMiddleware correctly deals with users
with valid tokens and those that have been deactivated, deleted or
replaced.

A unique ID provided by xid[0] is now appended to JWTs. This mitigates
the problem that, while testing, two tokens created in the same second
would match as they had an identical seed.

[0] https://github.com/rs/xid