Add authAllowCreateInitialUser endpoint
The authAllowCreateInitialUser returns `{"enabled": true}` if users exist
in the database and `{"enabled": false}` if they don't. It is designed
to allow the login page to check whether or not to display the initial
user screen which should only be used once, immediately after the app
is accessed for the first time.