Enhance make build step to build back and frontend
Now executing `make build` will first clean the static output directory
(`dist`), build the frontend and copy the output into the correct place,
compile the Go backend and use Hugo to compile any files in the
repository to be served from `/`.

The static output directory was changed from `build` to `dist` because
outputting to `build` interfered with make's `make build` command (it
claimed that there was nothing to build).

The static directory can now be configured using the `static` key in the
backend's config.