Add Synopsis, Tags and Version to FrontMatter
To reduce duplication the frontmatter form entries have been moved to
their own component which is included by `DocumentEdit` and
`DocumentNew`.

Tags and Version have also been introduced but not 'plugged in'. Version
is likely going to be changed to `string` (rather than `int`) to allow
for semantic versioning; `1.0.12` rather than simply `1.1`.