Fix bug causing routing failure in cucumber env
When performing the `make build` step, the `NODE_ENV` environment
variable wasn't being set to `production`, meaning it was using an
incorrect portion of config and causing the `GET
/auth/create_initial_user` to point at the dev environment on `:8080`
instead of Cucumber's on `:9095`. Adding this to the `Makefile` solved
the problem.

Also, add some features covering the basics of initial setup and first
user creation.