Make headless_chrome the default Capybara driver
It was previosuly failing because the test's setup (i.e. creating new
users via the API) was running before the server had begun listening;
there's a delay short enough to work with regular Chrome but not short
enough for headless Chrome's super-quick startup time.

Now we wait for the server to respond positively every 0.1 second for
ten attempts. All tests pass and a sensible error is thrown if there's
no server up and running when it times out.