First go at extracting commits with deltas
So we can take a peek at what actually happened in prior commits through
the web interface, the new `/api/commits/hash` endpoint returns a JSON
document containing the actual patch, the number of changes and probably
before too long, the commit message and committer information.

Currently the diff is just displayed in the new `Commit` component in
`<pre>` tags, but a JavaScript library aimed at diff formatting should
be a better option in the short term.