Add getFileHistory to retrieve commit history
getFileHistory should return all commits that affect the given file.
They are returned as a slice of `HistoricCommit`, which is like `Commit`
but with the additional `EntryID` field.

Tests ensure that the history is always returned in order and can be
limited to a certain number of entries by the `size` param