Remove signature information from NewCommit
All commit operations will be committed under the name of the logged-in
user, so passing the committer `name` and `email` in the `NewCommit`
object is now redundant. All commit-related methods accept a `user`
argument which, in the case of HTTP handlers, is populated during
authentication.

Every connection to endpoints that begin with 'api/' is authenticated
and the auth middleware now saves the current user and makes it
available to subsequent handlers via the request's Context[0], see
0c7994745 for more details.

To aid testing a new, minimal middleware is availble to the API tests,
accessed by creating a server with `createTestServerWithContext()`. All
it does is populates the current user with the one defined in
`apiTestUser()`.

[0] https://golang.org/pkg/context/