Move contents extraction to separate function
Extracting the contents of files prior to committing them to the
repoistory is fairly complex and it's cleaner for this behaviour to be
in a separate, testable function. This attempt still requires the
setting of a `Base64Encoded` flag which may not be necessary; it might
make more sense to base this purely on the file extension.

Alternatively:  there may be a more[0] elegant[1] solution

[0] https://mimesniff.spec.whatwg.org/
[1] https://golang.org/src/net/http/sniff.go