Remove hardcoded image handling
If the request URI matches an image type automatically try to serve the
image by extracting the relevant parts of the path (see
`extractImagePath`). This works in all cases **except** where the
document name isn't present.

For example, if the relative path was `images/screenshot.png` and the
preview URL was `documents/document-13.md`, the relative path (from
`documents/`) wouldn't include `document-13`.

A potential solution could be to use the `Referer` header to work out
which page we're on and find the correct file accordingly.

Another solution could be to, on page load, add an `onerror` function to
the images within the rendered Markdown document that re-requests them
and automatically prepends the document name to the `src` attribute