Make the editor's tags input more familiar
The tags field now uses the `tags-input` library[0] to support tag entry
in a normal fashion. Pressing enter will complete a tag, and it can be
deleted with backspace. A new feature ensures the library is being used
correctly but saving is not yet covered.

If not tags are added, `commit.js` will send an empty array.

[0] https://github.com/developit/tags-input