Make Delve/CodeLens work on tests
CodeLens[0] adds a 'debug test' option to each Go test, which uses Delve[1]
to debug the program. Unfortunately, VSCode's Go plugin[2] does not yet
support the setting of args or environment variables[3], which means we
cannot pass in our custom `-config` flag. As a result, the default
config location is used, `/etc/graphia.yml` and the test fails.

For the time being (until a better solution is found) I have switched
this to look at the test config file, `config/test.yml` and left a note
as to why.

[0] https://code.visualstudio.com/blogs/2017/02/12/code-lens-roundup
[1] https://github.com/derekparker/delve
[2] https://github.com/Microsoft/vscode-go/
[3] https://github.com/Microsoft/vscode-go/issues/879