Make setting tags more robust
Something slightly odd was happening in thebrowser with the `tags`
attribute of a ES6 class.

```javascript
document.tags    		=> ""
document.tags == ""             => true
document.tags === ""            => false
typeof document.tags            => 'object'
```

This caused a problem when form is submitted while the tags field is
empty, as `document.tags.split(",")` crashed (because `split` was being
called on a non-String).

Fixed by moving the conversion 'upstream' to the setter and actually
storing the array of tags in the `Document` class