Replace attachmentsDirectory with Slug everywhere
The FrontMatter's Slug attribute makes more sense because Hugo uses it
when publishing. If we set the slug to equal the filename (minus the
extension), the directory structure will be clean and sensible:

So, in the editor if we set the title of a new document in the documents
directory to "Once Upon a Time in Springfield", the frontmatter will
look like this:

```yaml
title: Once Upon a Time in Springfield
slug: once-upon-a-time-in-springfield
```

and the filename will be `once-upon-a-time-in-springfield.md`. Any
images dragged into the document will have a filepath like this:
`documents/once-upon-a-time-in-springfield/images/image-2.gif`, but the
markdown image tag will look like this:

```markdown
![image-2.gif](images/image-2.gif`)
```

Clean and simple.

The benefit of this approach is that when Hugo generates the output the
short image path will 'just work' because of the way Hugo generates
directories per document and publishes the document to `index.html` in
the associated directory.