Failed attempt at drag/drop image scenario
I just can't get it to work, more investigation required.

In theory, `source.drag_to(target)` should work, and appears to be
attempted by the driver **but** nothing actually happens. Perhaps the
target is wrong and it should be an element higher in the tree that
isn't the 'last line' 🤷