Move repo initialisation out of before all step
Until now there has been no need for cucumber features to run with an
empty directory where the git repo should be, so it was initialised in
the `Before` block.

Now we're testing initialisation this functionality has been moved to a
step which has been added to the `Background` of every feature that
needs it.