Rewrite document metadata using _index.md files
Hugo already has functionality that covers the requirements for document
metadata so `index.json` files are superfluous. Creating a `_index.md`
file, as suggested by Shadow53 in this thread[0], means that we can
leverage Hugo's existing functionality to load minimal Markdown files
that *only contain FrontMatter*, which can be encoded and decoded via
Particle in the regular fashion.

This simplifies both the backend logic and the theme.

[0] https://discourse.gohugo.io/t/accessing-the-contentdir-location-from-within-a-template/8271/5