Directory summary format improvements
Instead of returning a 'hash' type response, now an array of the new
type, DirectorySummary is returned. This allows metadata to be included
and displayed appropriately in the CMS. Titles and descriptions are
included automatically on the homepage.

Additionally, the new directory form has been split out into its own
component, ready for form validation and tests to be added.

Finally, getFilesInDir has been amended so it *does not return*
`_index.md` files. These will be handled differently to 'regular'
managed files.