Add Body property to DirectoryInfo
Hugo allows for the `_index.md` files to not only contain YAML but also
regular Markdown copy. It makes sense to use this for text that we might
wish to appear with the file listing but not necessarily in the summary
or homepage. This text is accessed in the theme/template using the
`{{ .Content }}` tag, so no more additional work is required.

The `body` attribute is ignored from the YAML[0] with `-`, so it won't
appear in the FrontMatter, but can be dealt with separately by
`writeMetadataFiles` via Particle.

[0] https://godoc.org/gopkg.in/yaml.v2#Marshal