Add apiGetDirectoryMetadata handler
So the breadcrumbs for pages that are beneath a directory but don't
actually load any data (i.e. the new document page), the directory info
is required to properly display the directory's index page breadcrumb.

If it's not present it will default to the directory's path, which isn't
capitalised, won't contain accents and won't nicely handle spaces