Make summary building much faster
The original method of building the summary, which scans every managed
markdown file and extracts/parses both the FrontMatter and the entire
contents of the file is quite slow when working with a either a small
number of very large files or a large number of smaller ones. This isn't
really what Particle[0] was designed for.

This modification replaces Particle with a custom file scanner that
looks for YAML FrontMatter at the top of a file and extracts only the
content between the FrontMatter boundaries '---':

```markdown
---
title: Marge vs. the Monorail
author: Conan O'Brien
---

Body goes here...
```

This method has one potential issue, if the file has an opening
FrontMatter boundary but no closing one, the whole file will be scanned
and passed to the YAML unmarshaller, which (like Particle), will reject
it and return an empty `FrontMatter{}` struct (see the tests). This
could potentially be a performance problem but I'm not sure there's a
failsafe way of detecting whether the line we're currently scanning is
YAML or not - mainly due to YAML's flexibility and ability to encode
multiple lines of text.

[0] https://github.com/njones/particle