Add RepositoryInfo type, endpoint and handler
The `RepositoryInfo` type contains data about the repository. Currently,
it has only the `LatestRevision` attribute, which is the string hash of
the latest commit. `RepositoryInfo` is also embedded into Files that are
sent to the client so the value can be included when sending back any
amendments.

This change will facilitate the checking of whether the served version
of the file is still up to date before writing to the repository.