Ensure RepositoryInfo is validated by API
The `RepositoryInfo` object, which contains `LatestRevision` must be
present when initiating a commit via the API. The validation directive
is present on the struct and this commit adds a test that ensures it
fails in the correct manner (with the right error message) should it be
absent.

The second part of the validation is that when we've validated that
`RepositoryInfo` is present, that the `LatestRevision` hash matches the
latest revision in the repository. This check is performed by
`checkLatestRevision`.

The final (unfinished) step is to ensure that if the repo is out of
date, a suitable error (identified as `ErrRepoOutOfSync`), is passed
back to the client.