Add the out of sync modal to new and edit screens
When a commit that modifies a file is attempted and fails with a
`Conflict` status (`409`), a modal is displayed. This will eventually
contain a full explanation of what's happend and have a button that
allows the user to download *their copy* of the file.

The modal is present and triggered by the `Conflict` status, but the
extra behaviour isn't yet present.