Add @chrome tag to download test and check content
Chrome Headless currently will not allow downloads to take place despite
the `DRIVER_PREFS` set in `env.rb`[0]. As a result, the one scenario
that relies on testing the content of a download is tagged as `@chrome`,
as with the regular Chrome driver it performs correctly. Omitting
`@chrome` from Cucumber's standard options means this test won't be run
unless explicitly set, but means that the full test suite will run
headlessly (which later might be important for CI).

[0] https://bugs.chromium.org/p/chromium/issues/detail?id=696481