Add the LanguageField to the editor and set ext
In Hugo, translations are denoted by the presence of a language code in
the filename[0]. So, the french version of `document.md` would be called
`document.fr.md`.

This is configured at file creation time with the choice of a language
from a dropdown on the editor form in creation mode, the language code
is automatically inserted into the filename.

[0] https://gohugo.io/content-management/multilingual/