Add NewTranslation object and createTranslation
NewTranslation works similarly to NewCommit but is much simpler. When a
source file is specified it copies it entirely to a new file but changes
the extension from `.md` to `.xy.md` where `xy` is the country code
specified in the config.

A commit message is generated automatically based on the language
selected.