Add and integrate the Translation component
The translation component is a dropdown that contains a list of the
languages enabled on the CMS. When selected, a new translation will be
initiated in that language. The actual process of
translation/localisation is outside the scope of this work, but a
XLIFF[0] download would be great. The md2xliff project[1] might be of
use here.

[0] http://docs.oasis-open.org/xliff/xliff-core/xliff-core.html
[1] https://github.com/cataria-rocks/md2xliff