Add SSH key fields to User and basic tests
The User model now has three additional fields:

* Type    - the type of SSH key (eg. RSA, ECDSA)
* Raw     - the key's raw content
* Comment - the helpful string on the end, usually email address

These are taken from the supplied string and stored without any parsing,
only a simple check is performed to make sure that the key is valid.

Storing the fully-parsed `ssh.PublicKey` was tricky due to the
implementation[0].

> PublicKey is an abstraction of different types of public keys.

BoltDB was reporting type mismatch errors between the public `PublicKey`
and the protected `rsaPublicKey`. Storing the raw data and parsing it in
an ad-hoc fashion is easier, it should be looked into more before this
work is merged though.

[0] https://godoc.org/golang.org/x/crypto/ssh#PublicKey