Add draft status to documents
Building on top of Hugo's draft functionality[0], being able to set
whether a document is or isn't a draft in the editor is a really useful
feature. It is implemented in a basic manner, with a simple checkbox
that is mapped to the `Draft` attribute of a `File`.

Draft files are highlighted with a 'warning' (orange in Bootstrap)
border on the index pages and they are labelled as 'Draft'.

An enhancement would be to have an actual 'Publish' button somewhere in
the UI that toggles the draft status of a file, that should now be
easily implementable.

[0] https://gohugo.io/getting-started/usage/#draft-future-and-expired-content