Add Date FrontMatter field to files and the editor
Dates are used by Hugo for sorting and organising posts and documents.
The date field defaults to 'today' and the dates are saved in ISO8601[0]
format (albeit as a string in the Markdown document).

Date formatting is performed by Fecha[1], a lighter alternative to
Moment that only does formatting and parsing (not transformation).

The standard JavaScript date formatting functionality[2] does not appear to
be finalised and doing some 'normal' things (like changing the separator
from '/' to '-') appears to be unsupported.

[0] https://en.wikipedia.org/wiki/ISO_8601
[1] https://github.com/taylorhakes/fecha
[2] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DateTimeFormat