Fix scroll behaviour and standardise on buttons
Now all the danger buttons actually do dangerous things (like
deletion) and primary is used for navigation and saving. Success will
probably be reserved for the publish button and confirmations. Cancel is
a grey 'secondary' button so as not to clash with danger.

Also, the page is scrolled to the top when navigation occurs and if
broadcast messages are displayed they are scrolled to. This makes it
more obvious that it's happened.