Bundle-ise the frontend
Many of the document-related paths have been 'improved' by the addition
of the `document` parameter. So, instead of `/cms/pokemon/pikachu.fr.md`
the new format is `/cms/pokemon/pikachu/fr`. The filename's removal
makes the bundles easier to deal with and to support the additional
routing flexibility, optional params are used[0]

[0] https://github.com/vuejs/vue-router/blob/dev/examples/route-matching/app.js#L17