Remove references to Frontmatter Slug
`Slug` is no longer required since the move to bundles and the Document
attribute. Leaving `Slug` in actually caused an additional problem in
the published output. URLs included it unnecessarily, so when slug was
present the generated path was `/pokemon/bulbasaur/bulbasaur/`, and when
removed it was (correctly) `/pokemon/bulbasaur`.