Remove GitRepoPath and simplify path resolution
The gitRepoPath config option was intended to make for a more flexible
setup than we now need. There are two repositories that are available to
the CMS, the `content` and the `theme`. Now we're just using
`config.Repisotory` instead of appending the supplied argument to the
`config.gitRepoPath` setting and then checking for the directory's
existence.